divided filledForms routing and prototypes routing into separate files, removed unecessary get from server
